refactor #81: DTO 추가 및 패키지 구조 수정
- CreateNotificationRequestDto 추가
- GetNotificationsRequestDto 추가
