+++ src/main/java/org/project/monewping/domain/notification/controller/NotificationController.java
@@ -1,24 +1,25 @@
 package org.project.monewping.domain.notification.controller;
 
-import java.time.Instant;
+import jakarta.validation.Valid;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.project.monewping.domain.notification.dto.CursorPageResponseNotificationDto;
 import org.project.monewping.domain.notification.dto.NotificationDto;
+import org.project.monewping.domain.notification.dto.request.CreateNotificationRequestDto;
+import org.project.monewping.domain.notification.dto.request.GetNotificationsRequestDto;
+import org.project.monewping.domain.notification.dto.response.CursorPageResponseNotificationDto;
 import org.project.monewping.domain.notification.service.NotificationService;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
@@ -32,11 +33,13 @@ public class NotificationController {
 
     @PostMapping
     public ResponseEntity<List<NotificationDto>> testCreate(
-        @RequestParam UUID userId,
-        @RequestParam UUID resourceId,
-        @RequestParam String resourceType
+        @ModelAttribute @Valid CreateNotificationRequestDto createNotificationRequestDto
     ) {
-        log.info("userId = {}, resourceId = {}, resourceType = {}", userId, resourceId, resourceType);
+        UUID userId = createNotificationRequestDto.userId();
+        UUID resourceId = createNotificationRequestDto.resourceId();
+        String resourceType = createNotificationRequestDto.resourceType();
+        log.info("userId = {}, resourceId = {}, resourceType = {}",userId, resourceId, resourceType);
+
         List<NotificationDto> notificationDtos = notificationService.create(userId, resourceId, resourceType);
 
         return ResponseEntity
@@ -47,21 +50,21 @@ public class NotificationController {
     /**
      * 커서 기반으로 사용자 알림 목록을 조회하는 엔드포인트입니다.
      *
-     * @param cursor 사용자에게 마지막으로 전달된 커서 토큰 (ISO-8601 문자열, 옵션)
-     * @param after  보조 커서(createdAt 필터)로 사용할 Instant (ISO-8601, 옵션)
-     * @param limit  한 페이지당 조회할 최대 알림 개수 (필수)
+     * @param getNotificationsRequestDto 조회 파라미터를 담은 DTO
      * @param userId 요청 헤더 {@code Monew-Request-User-ID}로 전달된 사용자 ID (필수)
      * @return 커서 페이지네이션 결과를 담은 {@link CursorPageResponseNotificationDto}를 HTTP 200 OK로 래핑하여 반환
      */
     @GetMapping
     public ResponseEntity<CursorPageResponseNotificationDto> getNotifications(
-        @RequestParam(value = "cursor", required = false) String cursor,
-        @RequestParam(value = "after", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) Instant after,
-        @RequestParam("limit") int limit,
+        @ModelAttribute @Valid GetNotificationsRequestDto getNotificationsRequestDto,
         @RequestHeader("Monew-Request-User-ID") UUID userId
     ) {
         CursorPageResponseNotificationDto cursorPageResponseNotificationDto =
-            notificationService.findNotifications(userId, cursor, after, limit);
+            notificationService.findNotifications(
+                userId,
+                getNotificationsRequestDto.cursor(),
+                getNotificationsRequestDto.after(),
+                getNotificationsRequestDto.limit());
 
         return ResponseEntity
             .status(HttpStatus.OK)
