+++ src/main/java/org/project/monewping/domain/notification/service/NotificationService.java
@@ -7,15 +7,22 @@ import jakarta.validation.constraints.Pattern;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
-import org.project.monewping.domain.notification.dto.CursorPageResponseNotificationDto;
+import org.project.monewping.domain.notification.dto.response.CursorPageResponseNotificationDto;
 import org.project.monewping.domain.notification.dto.NotificationDto;
 
 public interface NotificationService {
 
-    List<NotificationDto> create(@NotNull UUID userId, @NotNull UUID resourceId, @NotBlank @Pattern(regexp = "Article|Comment", message = "resourceType은 Article 또는 Comment만 허용됩니다.") String resourceType);
+    List<NotificationDto> create(
+        @NotNull UUID userId,
+        @NotNull UUID resourceId,
+        @NotBlank @Pattern(regexp = "Article|Comment", message = "resourceType은 Article 또는 Comment만 허용됩니다.") String resourceType
+    );
 
     CursorPageResponseNotificationDto findNotifications(
-        @NotNull UUID userId, String cursor, Instant after, @NotNull @Min(1) int limit
+        @NotNull UUID userId,
+        String cursor,
+        Instant after,
+        @NotNull @Min(1) int limit
     );
 
     void confirmAll(@NotNull UUID userId);
